odo/backend/requirements.txt
# FastAPI and Server
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
pymongo==4.7.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Validation & Settings
pydantic==2.7.1
pydantic-settings==2.2.1

# API Calls
requests==2.31.0

# File Handling for Uploads
python-multipart==0.0.9

# OCR for Receipts
pytesseract==0.3.10
Pillow==10.3.0

# Environment Variables
python-dotenv==1.0.1